requirements.txt
absl-py==1.4.0
attrs==23.1.0
cffi==1.15.1
contourpy==1.0.7
cycler==0.11.0
flatbuffers==23.5.26
fonttools==4.39.4
h5py==3.8.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
mediapipe==0.10.0
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
protobuf==3.20.3
pycparser==2.21
pyopencl==2023.1
PyOpenGL==3.1.7
pyparsing==3.0.9
PyQt6==6.5.1
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
pyserial==3.5
python-dateutil==2.8.2
pytools==2022.1.14
qt6-applications==6.5.0.2.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sounddevice==0.4.6
threadpoolctl==3.1.0
typing_extensions==4.6.3
